moved exceptions to the validation code
